==> notebooks/3.0-gr-building-predictive-models.ipynb <==
# Building Predictive Models

import pandas as pd
import os
import numpy as np


### import the pre-processed data


processed_data_path = os.path.join(os.path.pardir, 'data','processed')
train_file_path = os.path.join(processed_data_path, 'train.csv')
test_file_path = os.path.join(processed_data_path, 'test.csv')

train_df = pd.read_csv(train_file_path, index_col='PassengerId')
test_df = pd.read_csv(test_file_path, index_col='PassengerId')

train_df.info()

test_df.info()

### Data Preparation

# most machine learning algorithms expect numerical arrays, so let's set those up
# Please note we are ONLY working on the training data frame here.  
X = train_df.loc[:,'Age':].as_matrix().astype('float') #all rows and columns, except SURVIVED
y = train_df['Survived'].ravel() #ravel creates a flat 1-d array

print X.shape, y.shape

print(X.shape, y.shape)

# X this means 891 rows, and 32 colums/features, not including survived
# y this means 891 one dimensional array, one feature, just containing 0 or 1 for Survived


#Now split the training data
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0)
print(X_train.shape, y_train.shape)
print(X_test.shape, y_test.shape)

#We've just split the data into 80/20 
# Now let's see how many positive class values are in the train and test data
# Mean will give us the proportion of positives (i.e. 1 = SURVIVED)
print('Mean survival in the 80% training data: {0:.3f}'.format(np.mean(y_train)))
print('Mean survival in the 20% training test data: {0:.3f}'.format(np.mean(y_test)))

# We have similar proportions in the 80% and 20% splits.  This is good - this is what we want, 
# an even distribution
#
# Only 39% of data has positive cases, so 61% are going to be negative.  That's a little biased
# in terms of our source data.  This is known as class imbalance.

## check scikit-learn version

import sklearn
sklearn.__version__
# we need at least 0.19 onwards

## Create baseline model

# We want to know what our data precision and accuracy looks like BEFORE we start to do machine learning
from sklearn.dummy import DummyClassifier
# The dummy classifier gives you a measure of "baseline" performance--i.e. the success rate one should expect to achieve even if simply guessing.
# https://stackoverflow.com/questions/29441943/what-is-the-theorical-foundation-for-scikit-learn-dummy-classifier
#


#create a model
# most_frequent we want the model to output the majority class - i.e. 0/Not Survived in our case
model_dummy = DummyClassifier(strategy='most_frequent', random_state=0)
#
# Train the model using fit() with the input data, and the output data.  
model_dummy.fit(X_train, y_train)

# let's evaluate the performance.
# The model will predict the output against X_test and then it will compare it against the actual y_test
# The classification model will represent the model accuracy
print('Score for baseline mode : {0:.2f}'.format(model_dummy.score(X_test, y_test)))

# 61% baseline accuracy.  Without any algorithms, if you always predict 0/Not Survived, you will have a 61% success rate

### other performance metrics

from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score
# Method signatures are Actual Output, Predicited Output.  We can use the predict function on the input data to easily get the prediciton
print('Accuracy for baseline model: {0:.2f}'.format(accuracy_score(y_test, model_dummy.predict(X_test))))
print('Confusion Matrix for baseline model: \n {0}'.format(confusion_matrix(y_test, model_dummy.predict(X_test))))
# Our confusion matrix shows that we are always predicting 0 (Negative class)

print('Precision for baseline model: {0:.2f}'.format(precision_score(y_test, model_dummy.predict(X_test))))
print('Recall for baseline model: {0:.2f}'.format(recall_score(y_test, model_dummy.predict(X_test))))

## Submit to Kaggle


#convert the test data frame to a 2d matrix in float type
# test_df is the data frame for which we do not have the answers - Kaggle has the answers!
# we submit our prediction output, and Kaggle compares with actual output.
test_X = test_df.as_matrix().astype('float')

# Get predicitions from the data
predictions = model_dummy.predict(test_X)
# All our predictions will be 0/Not Survived - which should have a 61% success rate, right?

# Combine our predicitions with their passengerIds
# The file to submit to Kaggle is in format PassengerId : Survived 
# Build it in a pandas data frame from the index value of the test_df dataframe
df_submission = pd.DataFrame({'PassengerId': test_df.index,'Survived': predictions})

#Let's look at the content of the data frame
df_submission.head()

#save this data frame down to disk
submission_data_path = os.path.join(os.path.pardir, 'data','external')
submission_file_path = os.path.join(submission_data_path, '01_dummy.csv')

df_submission.to_csv(submission_file_path, index=False) #index=False so there's no extra index column in the file

# We want to create a function to do all this, since we'll be doing this again and again for subseqnet submissions
def get_submission_file(model, filename):
    #convert to matrix
    test_X = test_df.as_matrix().astype('float')
    #make prediciton
    predictions = model.predict(test_X)
    #create submission data frame
    df_submission = pd.DataFrame({'PassengerId': test_df.index,'Survived': predictions})
    #create submission file object
    submission_data_path = os.path.join(os.path.pardir, 'data','external')
    submission_file_path = os.path.join(submission_data_path, filename)
    #write to disk
    df_submission.to_csv(submission_file_path, index=False)

# get the submission file - we have to manully submitt this file.
get_submission_file(model_dummy, '01_dummy.csv')

## Logistic Regression Modelling

# import the function from scikit-learn
from sklearn.linear_model import LogisticRegression
# create the model
model_lr_1 = LogisticRegression(random_state=0)
# train the model
model_lr_1.fit(X_train, y_train)
# evaluate the model
print('Score for logistic regression v0.1: {0:.2f}'.format(model_lr_1.score(X_test, y_test)))

# The model preformance improved from 0.61 to 0.83  
# Our predictive model is therefore better than our baseline


### Performance metrics for the model

# Method signatures are Actual Output, Predicited Output.  We can use the predict function on the input data to easily get the prediciton
print('Accuracy for logistic regression v0.1 model: {0:.2f}'.format(accuracy_score(y_test, model_lr_1.predict(X_test))))
print('Confusion Matrix for logistic regression v0.1  model: \n {0}'.format(confusion_matrix(y_test, model_lr_1.predict(X_test))))
# Our confusion matrix shows that we are always predicting 0 (Negative class)
print('Precision for logistic regression v0.1  model: {0:.2f}'.format(precision_score(y_test, model_lr_1.predict(X_test))))
print('Recall for basellogistic regression v0.1 ine model: {0:.2f}'.format(recall_score(y_test, model_lr_1.predict(X_test))))

# Again, the metrics have also improved (precision and recall were both 0 in baseline)
# The point of training a logistical regression model is to find the optimal weights/coefficients b0 and b1
# How to get the coefficients?
model_lr_1.coef_

# These are the coefficients for each of the different input features.
# These coefficients are also known as model weights or Parameters.  So in the process of training a model, these 
# parameters are internally optimised during the training process.  This is how training works.
#
# If we wanted to know about the relative impact of input features on the outputs, then we can leverage these coefficients
# to get some useful inferences...  TBD.

## Second Kaggle Submission

#build file
get_submission_file(model_lr_1, '02_lr.csv')

# our ranking improved :D

# Hyperparameter optimisation

# create the base model, using default on the function
model_lr = LogisticRegression(random_state=0)

# we want to use the grid search function, so we can do multiple optimisation passes to find the best model
from sklearn.model_selection import GridSearchCV

# What are the various parameters we want to tweak in our model during our search for the best params?
params = {'C': [1.0,10.0,50.0, 100.0, 1000.0], 'penalty': ['l1','l2']} #L1, L2
# set the output model
crossValLF = GridSearchCV(model_lr, param_grid=params, cv=3) # use K-Fold 3 cross validation split on the data

# fit the data into the model
crossValLF.fit(X_train, y_train)

# What are the 'best params'?
crossValLF.best_params_ # property to return the best settings

print('Best Score: {0:.2f}'.format(crossValLF.best_score_))

# So it looks like we were able to improve our model a little using this method of tweaking the hyperparmas.
# This is about as best as logisitical regression can perform.
print('Score for logistic regression model v0.2: {0:.2f}'.format(crossValLF.score(X_test, y_test))) # run it against the test data

## submit to kaggle again

# get the submission file via the function again
get_submission_file(crossValLF, '03_lr.csv')

# manually submit this to kaggle
#
# Your Best Entry 
# You advanced 836 places on the leaderboard!
# Your submission scored 0.78468, which is an improvement of your previous score of 0.77990. Great job!

## Feature Standardisation and Normalisation

# scikit-learn provides good tools for this
from sklearn.preprocessing import MinMaxScaler, StandardScaler

### Feature Normalisation

# this is when we want to get the features we are using into the same scal
scaler = MinMaxScaler()
# create teh scaled object
X_train_scaled = scaler.fit_transform(X_train) #fit_transform combines fits the scaler with X_train, then scale the content

# now we have 'scaled' the output on the defaults for the min/max scaler function
# let's see what happens to the data
X_train_scaled[:,0].min(), X_train_scaled[:,0].max()

# This is as expected
# Now we have to scale the test data - as if we are training the model on normalised data,
# we should also test it against normalised data too
# We use different function, without the fit
X_test_scaled = scaler.transform(X_test)

## Feature Standardisation


scaler = StandardScaler()  # will now have 0 mean and 1 standard deviation
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

## Create  our model AFTER standardisation

# base model
model_lr = LogisticRegression(random_state=0)
params = {'C': [1.0,10.0,50.0, 100.0, 1000.0], 'penalty': ['l1','l2']} #L1, L2
# set the output model
crossValLF = GridSearchCV(model_lr, param_grid=params, cv=3) # use K-Fold 3 cross validation split on the data
crossValLF.fit(X_train_scaled, y_train)

# What is the best score now?
crossValLF.best_score_

print('Score for logistic regression model v0.2: {0:.2f}'.format(crossValLF.score(X_test_scaled, y_test))) # run it against the test data

# another minor enhancement!

## Model Persistence

# Our stakeholders are happy with the model to use it against real-time data
# we will use pickle library
import pickle
# create the file paths
model_file_path = os.path.join(os.path.pardir,'models','lr_model.pkl')
scaler_file_path = os.path.join(os.path.pardir,'models','lr_scaler.pkl') # we have a standard scaler object for feature standardisation so we can standardise new realtime inputs

# open the file to write
model_file_pickle = open(model_file_path,'wb') #binary write is required
scaler_file_pickle = open(scaler_file_path,'wb')

# persist the model and scaler to disk
pickle.dump(crossValLF, model_file_pickle)
pickle.dump(scaler, scaler_file_pickle)

# close the file handle
model_file_pickle.close()
scaler_file_pickle.close()
print(model_file_path)
print(scaler_file_path)

### Load the persisted files to see if they are doing what we expect

# open the file
model_file_pickle = open(model_file_path,'rb') # we need to use rb in python3
scaler_file_pickle = open(scaler_file_path,'rb')
# load them
crossValLF_loaded = pickle.load(model_file_pickle)
scaler_loaded = pickle.load(scaler_file_pickle)
#close the files they are no longer required
model_file_pickle.close()
scaler_file_pickle.close()

crossValLF_loaded # yep it's loaded ok

scaler_loaded # yes this has loaded too

# use these loaded objects to run aginst the test data set
# use scaler_loaded to transform the test data frame into a scaled data frame
X_test_scaled = scaler_loaded.transform(X_test)
# Now calculate teh score using the loaded trained model
print('Score for persisted logical regression v0.1: {0:.2f}'.format(crossValLF_loaded.score(X_test_scaled, y_test)))

# so we get the same score as we did on the fly previously
# So pickle is making it easy to store the model and the standardisation data to disk
